requirements.txt
asgiref==3.5.2
astroid==2.12.9
backports.zoneinfo==0.2.1
black==22.8.0
click==8.1.3
dill==0.3.5.1
dj-database-url==1.0.0
Django==4.1.1
django-filter==22.1
django-heroku==0.3.1
djangorestframework==3.13.1
Faker==14.2.0
gunicorn==20.0.4
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
mypy-extensions==0.4.3
pathspec==0.10.1
platformdirs==2.5.2
psycopg2==2.9.3
pylint==2.15.2
python-dateutil==2.8.2
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.4
typed-ast==1.5.4
typing-extensions==4.3.0
validate-docbr==1.10.0
whitenoise==6.2.0
wrapt==1.14.1
